requirements.txt
python-dotenv
slack-sdk
requests
